Map locales in separate file
