+++ src/components/map/MapApi.js
@@ -1,32 +1,6 @@
 import MapApi, { layerTypes, controlTypes } from '@dhis2/gis-api';
 // import MapApi, { layerTypes, controlTypes } from '@dhis2/maps-gl';
-import i18n from '@dhis2/d2-i18n';
-
-const getMapLocale = () => ({
-    acres: i18n.t('acres'),
-    Area: i18n.t('Area'),
-    Cancel: i18n.t('Cancel'),
-    'Center map on area': i18n.t('Center map on area'),
-    'Center map on line': i18n.t('Center map on line'),
-    'Click the next position': i18n.t('Click the next position'),
-    'Click where you want to start the measurement': i18n.t(
-        'Click where you want to start the measurement'
-    ),
-    Delete: i18n.t('Delete'),
-    Distance: i18n.t('Distance'),
-    'Enter fullscreen': i18n.t('Enter fullscreen'),
-    'Exit fullscreen': i18n.t('Exit fullscreen'),
-    ha: i18n.t('ha'),
-    km: i18n.t('km'),
-    'Measure distances and areas': i18n.t('Measure distances and areas'),
-    mi: i18n.t('mi'),
-    Perimeter: i18n.t('Perimeter'),
-    'Reset bearing to north': i18n.t('Reset bearing to north'),
-    'Search for place or address': i18n.t('Search for place or address'),
-    'Zoom in': i18n.t('Zoom in'),
-    'Zoom out': i18n.t('Zoom out'),
-    'Zoom to content': i18n.t('Zoom to content'),
-});
+import getMapLocale from './mapLocale';
 
 // Returns a new map instance
 const map = options => {
